+++ pyserini/search/faiss/__main__.py
@@ -20,16 +20,16 @@ import os
 import numpy as np
 from tqdm import tqdm
 
-from pyserini.encode import QueryEncoder
+from pyserini.encode import QueryEncoder, AutoQueryEncoder
 from pyserini.encode import AggretrieverQueryEncoder, AnceQueryEncoder, BprQueryEncoder, CosDprQueryEncoder, \
     DkrrDprQueryEncoder, DprQueryEncoder, TctColBertQueryEncoder
 from pyserini.encode._pca import PcaEncoder
 from pyserini.output_writer import get_output_writer, OutputFormat
 from pyserini.query_iterator import get_query_iterator, TopicsFormat
-from pyserini.search.faiss import FaissSearcher, BinaryDenseFaissSearcher
-from pyserini.search.faiss import DenseVectorAveragePrf, DenseVectorRocchioPrf, DenseVectorAncePrf
-from pyserini.search.faiss._searcher import AutoQueryEncoder, OpenAIQueryEncoder, ClipQueryEncoder
 from pyserini.search.lucene import LuceneSearcher
+from ._prf import DenseVectorAveragePrf, DenseVectorRocchioPrf, DenseVectorAncePrf
+from ._searcher import OpenAIQueryEncoder, ClipQueryEncoder
+from ._searcher import FaissSearcher, BinaryDenseFaissSearcher
 
 # Fixes this error: "OMP: Error #15: Initializing libomp.a, but found libomp.dylib already initialized."
 # https://stackoverflow.com/questions/53014306/error-15-initializing-libiomp5-dylib-but-found-libiomp5-dylib-already-initial
